+++ project-base/storefront/hooks/app/useAuthLoader.ts
@@ -1,15 +1,32 @@
+import { getCookies } from 'cookies-next';
 import { isClient } from 'helpers/isClient';
+import { getUrlWithoutGetParameters } from 'helpers/parsing/urlParsing';
 import { showInfoMessage, showSuccessMessage } from 'helpers/toasts';
+import { useIsUserLoggedIn } from 'hooks/auth/useIsUserLoggedIn';
 import useTranslation from 'next-translate/useTranslation';
+import { useRouter } from 'next/router';
 import { useEffect } from 'react';
 import { usePersistStore } from 'store/usePersistStore';
 
 export const useAuthLoader = () => {
     const { t } = useTranslation();
     const authLoading = usePersistStore((store) => store.authLoading);
+
+    const router = useRouter();
+    const isUserLoggedIn = useIsUserLoggedIn();
+    const slug = getUrlWithoutGetParameters(router.asPath);
     const updateAuthLoadingState = usePersistStore((store) => store.updateAuthLoadingState);
 
     const isStoreHydrated = isClient && usePersistStore.persist.hasHydrated();
+
+    useEffect(() => {
+        const cookies = getCookies();
+        const isWithUserTokens = !!(cookies.accessToken && cookies.refreshToken);
+
+        if ((isUserLoggedIn && !isWithUserTokens) || (!isUserLoggedIn && isWithUserTokens)) {
+            router.reload();
+        }
+    }, [slug]);
 
     useEffect(() => {
         if (authLoading && isStoreHydrated) {
