+++ project-base/storefront/hooks/cart/useAddToCart.ts
@@ -6,6 +6,7 @@ import { GtmMessageOriginType, GtmProductListNameType } from 'gtm/types/enums';
 import { mapPriceForCalculations } from 'helpers/mappers/price';
 import { showErrorMessage } from 'helpers/toasts';
 import { useIsUserLoggedIn } from 'hooks/auth/useIsUserLoggedIn';
+import { dispatchBroadcastChannel } from 'hooks/useBroadcastChannel';
 import { useDomainConfig } from 'hooks/useDomainConfig';
 import useTranslation from 'next-translate/useTranslation';
 import { usePersistStore } from 'store/usePersistStore';
@@ -53,6 +54,8 @@ export const useAddToCart = (
         if (!addToCartResult) {
             return null;
         }
+
+        dispatchBroadcastChannel('refetchCart');
 
         const addedCartItem = addToCartResult.addProductResult.cartItem;
         const notOnStockQuantity = addToCartResult.addProductResult.notOnStockQuantity;
